This automation script is very powerful! Use with EXTREAM care!
Look, this is a ML script that will not work without my trained ML model to go along with it, but it might... I am not sure. The payloads alone are the most powerful part and were made by jailbreaking multiple AI models (perplexity, ChatGPT, Claude, etc.) to craft and recreate the most powerful payloads for SQLi, SSRF, XXE, XSS, and RCE. These are the most updated payloads used/made in 2023 - 2024 and some I made using the jailbroken AI models. 

For the ML part, I used sCapy library combined with other modules in python to make it, so this trained model looks for all the key signals of there being a vuln/bug in the area and logs what it found with detailed data returned in files that are 10MBs each. So, the data returned is all the data needed to confirm the bug is really present and is where the model says it is. This is very accurate when the payloads, user-agents, wordlist and other parts of this script all come together. 

This is the most powerful ML bug hunting python automation script for hacking I have ever made and is beyond good at what it does. 

If you have any questions or would like to add to the script, ask me first, please!